refactor(all): refactor several aspects
- renamed scrape.py to utils.py
- cleaned views.py from unnecessary comments and unused imports
- removed print_hello, print_task and scrape_interval_task_old tasks
- deleted the views that corresponded to the removed tasks above
- deleted url patterns that were left over from deleted views
- cleaned up celery.py from unnecessary comments
