+++ main/views.py
@@ -1,7 +1,5 @@
 import re
-from datetime import timedelta, datetime
 
-from celery import current_app
 from django.contrib.auth import get_user_model
 from django.shortcuts import render, redirect
 from django.views.generic import ListView, DetailView
@@ -9,11 +7,9 @@ from guardian.mixins import PermissionListMixin, PermissionRequiredMixin
 
 from django_celery_beat.models import PeriodicTask, IntervalSchedule
 
-from mp_monitor.celery import app
-from .forms import ScrapeForm, TaskForm, ScrapeIntervalForm
-from .models import Item, Price, Printer
-from .scrape import get_scraped_data
-from .tasks import print_task
+from .forms import ScrapeForm, ScrapeIntervalForm
+from .models import Item, Price
+from .utils import get_scraped_data
 
 from django.utils import timezone
 
@@ -63,6 +59,7 @@ class ItemDetailView(PermissionRequiredMixin, DetailView):
                 else:
                     prices[i].table_class = 'table-warning'
                     prices[i].trend = '='
+
             # the original price is the last price in the list, so no comparison is possible
             except IndexError:
                 prices[i].table_class = ''
@@ -100,90 +97,11 @@ def scrape_items(request, skus):
     return render(request, "main/item_list.html", {"form": form})
 
 
-def trigger_print_task(request):
-    task_date = datetime.now().strftime('%Y-%m-%d_%H:%M:%S')
-    # task = 'hello'
-    printers =  Printer.objects.all()
-    if request.method == "POST":
-        task_form = TaskForm(request.POST)
-        if task_form.is_valid():
-            interval = task_form.cleaned_data["interval"]
-
-            schedule, created = IntervalSchedule.objects.get_or_create(
-            every = interval,
-            period = IntervalSchedule.SECONDS,
-            )
-
-            task = PeriodicTask.objects.create(
-                interval=schedule,
-                name=f'task_{request.user}',
-                task='main.tasks.print_task',
-                # expires=datetime.now() + timedelta(seconds=30)
-            )
-            # task.save()
-
-            # store 'task' in session to display as context in print_task.html
-            request.session['task'] = f'{task.name} - {task.interval}'
-
-            # To stop task
-            # periodic_task = PeriodicTask.objects.get(name='task_name')
-            # periodic_task.enabled = False
-            # periodic_task.save()
-
-
-                # print_task.apply_async(countdown=interval)
-
-            # task = current_app.send_task('main.tasks.print_task')
-            # print_task.apply_async(countdown=interval, repeat=True)
-            # #
-
-            return redirect("print_task")
-
-            # context = {
-            #     'task_form': task_form,
-            #     'printers': printers,
-            #     'task': task,
-            # }
-            # return render(request, "main/print_task.html", context)
-    else:
-        task_form = TaskForm()
-
-    context = {
-        'task_form': task_form,
-        'printers': printers,
-        'task':  request.session.get('task'),
-    }
-    return render(request, "main/print_task.html", context)
-
-def destroy_print_task(request):
-    # if request.method == "POST":
-    #     task_form = TaskForm(request.POST)
-    #     if task_form.is_valid():
-    # interval = task_form.cleaned_data["interval"]
-
-    periodic_task = PeriodicTask.objects.get(name=f'task_{request.user}')
-    # periodic_task.enabled = False
-    # periodic_task.save()
-    periodic_task.delete()
-    print(f"{periodic_task} has been deleted!")
-
-    # delete 'task' from session
-    if 'task' in request.session:
-        del request.session['task']
-
-    return redirect("print_task")
-
-
-def scrape_interval_task(request):
+def create_scrape_interval_task(request):
     """
     Takes interval from the form data (in seconds) and triggers main.tasks.scrape_interval_task
     The task itself prints all items belonging to this tenant every {{ interval }} seconds.
     """
-    # task_date = datetime.now().strftime('%Y-%m-%d_%H:%M:%S')
-    # # task = 'hello'
-    # items = Item.objects.filter(tenant=request.user.tenant)
-    # items_skus = [item.sku for item in items]
-    # print(f"{items_skus=}")
 
     if request.method == "POST":
         scrape_interval_form = ScrapeIntervalForm(request.POST)
@@ -191,19 +109,19 @@ def scrape_interval_task(request):
             interval = scrape_interval_form.cleaned_data["interval"]
 
             schedule, created = IntervalSchedule.objects.get_or_create(
-            every = interval,
-            period = IntervalSchedule.SECONDS,
+                every=interval,
+                period=IntervalSchedule.SECONDS,
             )
 
             scrape_interval_task = PeriodicTask.objects.create(
                 interval=schedule,
                 name=f'scrape_interval_task_{request.user}',
                 task='main.tasks.scrape_interval_task',
-                start_time=timezone.now(), # trigger once right away and then keep the interval
+                start_time=timezone.now(),  # trigger once right away and then keep the interval
                 args=[request.user.tenant.id],
             )
 
-            # store 'scrape_interval_task' in session to display as context in print_task.html
+            # store 'scrape_interval_task' in session to display as context in item_list.html
             request.session['scrape_interval_task'] = f'{scrape_interval_task.name} - {scrape_interval_task.interval}'
 
             return redirect("item_list")
@@ -213,36 +131,18 @@ def scrape_interval_task(request):
 
     context = {
         'scrape_interval_form': scrape_interval_form,
-        'scrape_interval_task':  request.session.get('scrape_interval_task'),
+        'scrape_interval_task': request.session.get('scrape_interval_task'),
     }
     return render(request, "main/item_list.html", context)
 
-def destroy_scrape_interval_task(request):
-    # if request.method == "POST":
-    #     task_form = TaskForm(request.POST)
-    #     if task_form.is_valid():
-    # interval = task_form.cleaned_data["interval"]
 
+def destroy_scrape_interval_task(request):
     periodic_task = PeriodicTask.objects.get(name=f'scrape_interval_task_{request.user}')
-    # periodic_task.enabled = False
-    # periodic_task.save()
     periodic_task.delete()
     print(f"{periodic_task} has been deleted!")
 
-    # delete 'task' from session
+    # delete 'scrape_interval_task' from session
     if 'scrape_interval_task' in request.session:
         del request.session['scrape_interval_task']
 
     return redirect("item_list")
-
-
-
-
-
-# TODO: think how to scrape category
-# TODO: think how to scrape image
-# TODO: remove seller_name since it's the same as brand?
-# TODO: add scrape_item to ListView's post method, or create a separate ItemCreateView:
-
-# CreateView?
-# https://github.com/django-guardian/django-guardian/blob/55beb9893310b243cbd6f578f9665c3e7c76bf96/example_project/articles/views.py#L19C1-L30C20
